+++ tests/lib.rs
@@ -8,7 +8,7 @@ use loveletter::Play;
 use loveletter::PlayError;
 
 
-fn make_round(hands: &[Option<Card>], deck: &[Card], current_player: Option<uint>) -> (loveletter::Round, Vec<loveletter::PlayerId>) {
+fn make_round(hands: &[Option<Card>], deck: &[Card], current_player: Option<usize>) -> (loveletter::Round, Vec<loveletter::PlayerId>) {
     let game = loveletter::game::new_game(hands.len()).unwrap();
     let players = game.players();
     assert_eq!(players.len(), hands.len());
